// src/features/vehicles/admin/components/EditVehicle/types.ts
export type EditVehicleValues = {
  codeFasecolda: string
  licencePlate: string
  modelId: string
  serviceId: number
  brandId: string
  chassisNumber: string
  originId: number
  classId: string
  engine: string
  lineId: number | null
  mainColor: string
  secondaryColor: string
  serie: string
  vin: string
  conditionId: number
  grossValue: number
  doors: number
  capacity: number
  statusVehicleId: number
  typeFuelId: number
  officeId: number
  typeBodyworkId: number
  mileage: string
  description: string
  typeTransmissionId: number
  typeTractionId: number
  dateEntry: string
}